prizeEa is float

+++ Domain/Product.cs
@@ -6,7 +6,7 @@ namespace Domain
     {
         public Guid Id { get; set; }
         public string ProductName { get; set; }
-        public int PriceEa { get; set; }
+        public float PriceEa { get; set; }
         public int NumberOfProducts { get; set; }
     }
 }
